% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMtGeneLabel.R
\name{addMtGeneLabel}
\alias{addMtGeneLabel}
\title{Set gene names for bp ranges. Original code by Stephen Turner at UVA.}
\usage{
addMtGeneLabel(mtCalls)
}
\arguments{
\item{mtCalls}{a VRanges of mitochondrial calls with VAF and PASS mcols}
}
\value{
a coordinate-sorted VRanges with additional annotations
}
\description{
FIXME: this is only for hg19/GRCh37 at the moment, need to add hg38/GRCh38
}
