% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_rCRS.R
\docType{data}
\name{anno_rCRS}
\alias{anno_rCRS}
\title{annotation for the regions (genic and not) of the rCRS mitogenome}
\format{
a GRanges object
}
\usage{
anno_rCRS
}
\description{
annotation for the regions (genic and not) of the rCRS mitogenome
}
\examples{
data(anno_rCRS)
subset(anno_rCRS, region == "D-loop")

}
\keyword{datasets}
