% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byStrand.R
\name{byStrand}
\alias{byStrand}
\title{simple helper function to split a *RangesList by the strand of its mt target}
\usage{
byStrand(x, anno = NULL)
}
\arguments{
\item{x}{a *Ranges[List] or *Alignments[List]}

\item{anno}{optional feature annotation, will use mtAnno.rCRS if NULL}
}
\value{
elements of x over features on each strand OR x split by strand
}
\description{
If presented with a GAlignments/MAlignments, this method will split the 
element by strand, i.e. + alignments and - alignments. Otherwise the method
retrieves ranges/variant calls that overlap genic elements on the heavy and
light strands of the mitochondrial genome.
}
\examples{
data(RONKSvariants, package="MTseekerData")
byStrand(RONKSvariants)

}
