% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callTNMT.R
\name{callTNMT}
\alias{callTNMT}
\alias{callTnMtVars}
\title{call mitochondrial tumor/normal variants from an MAlignments[List] object}
\usage{
callTNMT(tmal, nmal, ...)

callTnMtVars(BAMS, SIZE, COV, GENOME = "rCRS", CHR = "chrM")
}
\arguments{
\item{tmal}{tumor MAlignments}

\item{nmal}{normal MAlignments}

\item{...}{other optional arguments to pass to callVariants}

\item{BAMS}{the BAM filenames}

\item{SIZE}{the read length}

\item{COV}{the average read coverage}

\item{GENOME}{the reference genome (default rCRS)}

\item{CHR}{the mt contig name (default is chrM)}
}
\description{
call mitochondrial tumor/normal variants from an MAlignments[List] object
}
