% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrominfo.rCRS.R
\docType{data}
\name{chrominfo.rCRS}
\alias{chrominfo.rCRS}
\title{Sequence information (seqinfo) for the rCRS mitogenome}
\format{a Seqinfo object}
\usage{
chrominfo.rCRS
}
\description{
Sequence information (seqinfo) for the rCRS mitogenome
}
\examples{
data(chrominfo.rCRS)
library(GenomeInfoDb) # wat
seqlengths(chrominfo.rCRS) 

}
\keyword{datasets}
