% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMAlignments.R
\name{combineMAlignments}
\alias{combineMAlignments}
\title{Combine multiple MAlignments objects together}
\usage{
combineMAlignments(mall, malToCombine = NULL, replace = FALSE)
}
\arguments{
\item{mall}{An MAlignmentsList object}

\item{malToCombine}{MAlignments object names to combine}

\item{replace}{Whether to replace the merged object into the MAlignmentsList in place}
}
\value{
Either a combined MAlignments object or an updated MAlignmentsList
}
\description{
Combine multiple MAlignments objects together
}
\examples{

library(MTseeker)
library(MTseekerData)
snames <- names(RONKSreads)[grepl("NK", names(RONKSreads))][1:3]
mergedMAlist <- combineMAlignments(RONKSreads, malToCombine=snames, replace=TRUE)

}
