% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeAndCalcConsequences.R
\name{decomposeAndCalcConsequences}
\alias{decomposeAndCalcConsequences}
\title{Decompose and annotate AA changes in MT variants}
\usage{
decomposeAndCalcConsequences(mvr, coding = TRUE, AAchanges = TRUE, ...)
}
\arguments{
\item{mvr}{An MVRangesList or MVRanges object}

\item{coding}{Annotate only coding regions? (TRUE)}

\item{AAchanges}{Annotate amino acid (AA) changes? (TRUE)}

\item{...}{Other arguments to pass to injectMTVariants}
}
\value{
\preformatted{       Annotated variants, as an MVRanges
}
}
\description{
FIXME: this function is remarkably slow
}
\examples{
library(MTseeker)
library(MTseekerData)
library(VariantTools)

# Set a really high depth filter
# Just an example, not something you'd use to filter real data
# Something like 20 reads is more reasonable (gives Sanger-like error rates)
filters <- 
  FilterRules(list(minTotalDepth=MinTotalDepthFilter(min.depth=2e3L)))
ronks_vars.anno <- subsetByFilter(RONKSvariants[[1]], filters)
decomposeAndCalcConsequences(ronks_vars.anno)

}
