% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthFilterAndHaploMask.R
\name{depthFilterAndHaploMask}
\alias{depthFilterAndHaploMask}
\title{process an MVRangesList for 20x depth and haplomask it
note that any elements with < 1 variant after filtering are tossed}
\usage{
depthFilterAndHaploMask(MVRL, minDepth = 20)
}
\arguments{
\item{MVRL}{An MVRangesList}

\item{minDepth}{minimum depth to retain (20)}
}
\description{
process an MVRangesList for 20x depth and haplomask it
note that any elements with < 1 variant after filtering are tossed
}
