% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFilteredAsVcf.R
\name{exportFilteredAsVcf}
\alias{exportFilteredAsVcf}
\title{convenience wrapper for dumping filtered variants directly to an indexed VCF}
\usage{
exportFilteredAsVcf(vars, file = NULL, ...)
}
\arguments{
\item{vars}{an MVRanges or MVRangesList to filter, HGVS-annotate, and dump}

\item{file}{the name of the resulting VCF file to produce (default is NULL)}

\item{...}{other arguments, passed to filterMTvars()}
}
\value{
the filtered set of variants (as an MVRanges) that was exported
}
\description{
convenience wrapper for dumping filtered variants directly to an indexed VCF
}
\examples{
library(MTseekerData) 
filterMTvars(RONKSvariants$RO_1)
exportFilteredAsVcf(RONKSvariants$RO_1, file=NULL) # not written to a file 

}
