% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMT.R
\name{filterMT}
\alias{filterMT}
\title{Filter SummarizedExperiment or DataFrame values based on its $mtCovg column}
\usage{
filterMT(DFSE, minCovg = 20)
}
\arguments{
\item{DFSE}{a DataFrame/SummarizedExperiment with colData() $mtCovg}

\item{minCovg}{minimum covg (20, cf. Griffin, Genetics in Medicine 2014)}
}
\description{
Griffin et al. (Genetics in Medicine 2014) recommends 20x coverage for mtDNA
sequencing to have comparable error rates to Sanger sequencing.  By default, 
that is the cutoff applied here to ensure halfway decent variant annotation.
}
