% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMTvars.R
\name{filterMTvars}
\alias{filterMTvars}
\title{sanitize PASSing mitochondrial variant calls to a moderate degree}
\usage{
filterMTvars(vars, fp = TRUE, NuMT = 0.03, covg = 20, depth = 2)
}
\arguments{
\item{vars}{an MVRanges or MVRangesList (will be unlisted and relisted)}

\item{fp}{use false positive filter\link{s}? (TRUE: use Triska fpFilter)}

\item{NuMT}{variants with VAF < \link{this number} will be presumed NuMTs (0.03)}

\item{covg}{minimum median read coverage across chrM to be considered (20)}

\item{depth}{minimum altDepth for a variant to be retained (2)}
}
\value{
a filtered set of variants
}
\description{
sanitize PASSing mitochondrial variant calls to a moderate degree
}
\examples{
library(MTseekerData) 
filterMTvars(RONKSvariants$RO_1)

}
