% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMetadata.R
\name{fixMetadata}
\alias{fixMetadata}
\title{fix metadata for an MAlignmentsList or MVRangesList, if needed and possible.}
\usage{
fixMetadata(x)
}
\arguments{
\item{x}{an MAlignmentsList or MVRangesList}
}
\value{
the object, with fixed metadata (if needed and possible)
}
\description{
This function is punishingly simple -- it just calls validMetadata() and
assigns attr('fixedMeta') from the result. If no fixing is required, the 
object's existing metadata is used. The object (with fixed metadata) is 
then returned. If automatic fixes are impossible, a message is generated.
}
\examples{
library(MTseekerData)
data(RONKSreads)
fixed <- fixMetadata(RONKSreads)

}
