% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpFilter_RSRS.R
\docType{data}
\name{fpFilter_RSRS}
\alias{fpFilter_RSRS}
\title{false positive filter (fpFilter) for RSRS haplogroup-determining variants}
\format{a GRanges object}
\usage{
fpFilter_RSRS
}
\description{
false positive filter (fpFilter) for RSRS haplogroup-determining variants
}
\examples{
data(fpFilter_RSRS)
subset(fpFilter_RSRS, !is.na(L0))
subset(fpFilter_RSRS, !is.na(rCRS))

}
\keyword{datasets}
