% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpFilter_Triska.R
\docType{data}
\name{fpFilter_Triska}
\alias{fpFilter_Triska}
\title{a false positive (fp) filter from Petr Triska's manuscript on mtDNA variants}
\format{a GRanges object}
\usage{
fpFilter_Triska
}
\description{
a false positive (fp) filter from Petr Triska's manuscript on mtDNA variants
}
\examples{
data(fpFilter_Triska) 
show(fpFilter_Triska) 

}
\keyword{datasets}
