% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMTcircos.R
\name{genesMTcircos}
\alias{genesMTcircos}
\title{Actually plot the rings with the gene names}
\usage{
genesMTcircos(x, anno, legends = F)
}
\arguments{
\item{x}{something with a genome()}

\item{anno}{a GRanges with annotations}

\item{legends}{plot legends? (FALSE)}
}
\value{
\preformatted{   annotations
}
}
\description{
Actually plot the rings with the gene names
}
