% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMTcircos.R
\name{genMTcircos}
\alias{genMTcircos}
\title{Plot an empty circos plot of the mitochondrial genome}
\usage{
genMTcircos(mvr)
}
\arguments{
\item{mvr}{MVRanges}
}
\value{
\preformatted{   Image of mitochondrial genome
}
}
\description{
Plot an empty circos plot of the mitochondrial genome
}
