% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureCounts.R
\name{getFeatureCounts}
\alias{getFeatureCounts}
\title{utility fn for getComplexity and preseqR}
\usage{
getFeatureCounts(x)
}
\arguments{
\item{x}{the thing to count features in (usually a GRanges)}
}
\value{
an array of feature frequencies
}
\description{
utility fn for getComplexity and preseqR
}
