% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMT.R
\name{getMT}
\alias{getMT}
\title{grab the mitochondrial reads from a BAM & estimate their fraction (of total)}
\usage{
getMT(bam, filter = FALSE, parallel = FALSE, plotMAPQ = FALSE, ...)
}
\arguments{
\item{bam}{a BAM filename, or DataFrame/SummarizedExperiment with $BAM}

\item{filter}{filter on bam$mtCovg? (default is FALSE, don't filter)}

\item{parallel}{load multiple BAMs in parallel, if possible? (FALSE)}

\item{plotMAPQ}{plot distribution of mitochondrial mapping quality? (FALSE)}

\item{...}{additional args to pass scanBamParam(), such as mapqFilter}
}
\value{
an MAlignments or MAlignmentsList object
}
\description{
nb. this could probably be done faster for a list of BAMs but it's not
nb. nb. this returns NuMT-depleted mitochondrial GenomicAlignments
nb. nb. nb. for the time being, this function ONLY supports rCRS/GRCh/hg38!
nb. nb. nb. nb. both chrM and mtGenome are now autodetected from BAM headers
nb. nb. nb. nb. nb. in the process of converting to rCRS, chrM becomes "chrM"
}
\examples{
library(MTseekerData)
data(RONKSreads)
\donttest{
  RONKSreads <- getMT(BAMs)
} 
RONKSreads

}
