% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProteinImpact.R
\name{getProteinImpact}
\alias{getProteinImpact}
\title{Annotate AA changes in MT variants}
\usage{
getProteinImpact(mvr, parallel = FALSE, cores = 1)
}
\arguments{
\item{mvr}{An MVRangesList or MVRanges object}

\item{parallel}{Whether to run things in parallel}
}
\value{
Annotated variants
}
\description{
Annotate AA changes in MT variants
}
\examples{


}
