% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScReads.R
\name{getScReads}
\alias{getScReads}
\title{merge into MTseeker; could also do this with SmartSeq, etc.}
\usage{
getScReads(
  BAM,
  cellTag = "CB",
  fragTag = "UB",
  tagFilt = NULL,
  dedupe = FALSE,
  MTonly = FALSE,
  byCell = FALSE,
  byFrag = FALSE
)
}
\arguments{
\item{BAM}{a BAM file or a BamFile object}

\item{cellTag}{what tag to use for cell identification (CB)}

\item{fragTag}{what tag to use for fragment identification (UB)}

\item{tagFilt}{optional tag filter (eg. list(CB="ATCGATCGAT-1")) (NULL)}

\item{dedupe}{drop reads with missing or duplicate (CB,UB)? (FALSE)}

\item{MTonly}{only reaad in mitochondrial reads? (FALSE)}

\item{byCell}{split by cell? (FALSE; only set TRUE for MT)}

\item{byFrag}{split by fragment? (FALSE; only set TRUE for MT)}
}
\value{
\preformatted{         a GAlignments or GAlignmentsList (if byCell/byFrag)
}
}
\description{
For 10X BAMs, cellTag is CB and fragTag is UB.
For Salmon BAMs, cellTag is CB and fragTag is UR.
For BioRad BAMs, cellTag is DB and fragTag is XB (I think).
Per-cell pileups can be generated directly from BAMs, though this is tricky.
I'm not 100 percent sure how to deal with UMI reconciliation above, though.
}
