% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplomask_whitelist.R
\docType{data}
\name{haplomask_whitelist}
\alias{haplomask_whitelist}
\title{a whitelist for haplomasking}
\format{
an MVRangesList object
}
\usage{
haplomask_whitelist
}
\description{
a whitelist for haplomasking
}
\examples{
data(haplomask_whitelist) 
show(haplomask_whitelist) 

}
\keyword{datasets}
