% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19TorCRS.R
\docType{data}
\name{hg19TorCRS}
\alias{hg19TorCRS}
\title{a liftOver chain for hg19 to rCRS attempts. We do not recommend this.}
\format{
a Chain object
}
\usage{
hg19TorCRS
}
\description{
a liftOver chain for hg19 to rCRS attempts. We do not recommend this.
}
\examples{
data(hg19TorCRS) 
sapply(hg19TorCRS, offset)

}
\keyword{datasets}
