% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVCF.R
\name{importVCF}
\alias{importVCF}
\title{Import a VCF directly into MTseeker}
\usage{
importVCF(x, gzip = TRUE, biscuit = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{character, list or dataframe for file name of VCF(and BAM)}

\item{gzip}{Is the input gzipped?}

\item{biscuit}{Is the input VCF from biscuit?}

\item{verbose}{logical, if set to TRUE, print all messages during progressing.}
}
\value{
an MVRangesList
}
\description{
Import a VCF directly into MTseeker
}
\examples{


}
