% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexMtGenome.R
\name{indexMtGenome}
\alias{indexMtGenome}
\title{build and install GmapGenome.Hsapiens.[mtGenome], currently for rCRS}
\usage{
indexMtGenome(mtGenome = "rCRS", chrM = "chrM", destDir = NULL,
  install = TRUE)
}
\arguments{
\item{mtGenome}{mitochondrial reference genome to index (default is rCRS)}

\item{chrM}{what the mitochondrial contig is called (default is chrM)}

\item{destDir}{optional destination for the package ($HOME is default)}

\item{install}{install the package after creation? (default is TRUE)}
}
\value{
the path to the created package as a character string
}
\description{
gmapR needs a reference genome index against which to call variants.
We support (only) rCRS as that reference (at least for the time being). 
This function creates and installs a reference Gmap index, by default rCRS.
In principle, hg19 and mm10 are supported, but in practice, support is poor.
}
