% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMTcircos.R
\name{initMTcircos}
\alias{initMTcircos}
\title{Initialize the circos plots}
\usage{
initMTcircos(x)
}
\arguments{
\item{x}{something with a genome()}
}
\value{
\preformatted{   annotations
}
}
\description{
Initialize the circos plots
}
