% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectMTvariants.R
\name{injectMTVariants}
\alias{injectMTVariants}
\title{Inject (one or more) variants against rCRS.}
\usage{
injectMTVariants(mvr, coding, gr = NULL)
}
\arguments{
\item{mvr}{An MVRanges, usually from callMT, often subsetted}

\item{coding}{TRUE to only look at coding regions}

\item{gr}{A GRanges, usually of protein-n regions (the default)}
}
\value{
\preformatted{     The GRanges, with ref/var DNA and AA and 
}
}
\description{
FIXME: this function could most likely be orders of magnitude faster.
FIXME: this ONLY considers variants injected against rCRS, not RSRS or hg19.
}
\examples{
\dontrun{
library(MTseekerData)
RO_2 <- RONKSvariants[["RO_2"]]
injectMTVariants(RO_2)
}
}
