% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_up_to_rCRS.R
\name{lift_up_to_rCRS}
\alias{lift_up_to_rCRS}
\title{Lift hg19 chrM to rCRS}
\usage{
lift_up_to_rCRS(ranges_obj, verbose = T)
}
\arguments{
\item{ranges_obj}{Either GRanges or MVRanges}

\item{verbose}{logical, if set to TRUE, print all messages during progressing.}
}
\value{
\preformatted{      Either Granges or MVRanges object
}
}
\description{
Lift hg19 chrM to rCRS
}
\examples{

}
