% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateMTvariants.R
\name{locateMTvariants}
\alias{locateMTvariants}
\title{Locates the variant and determines local bounds within gene
Primarily meant to be used within injectMTvarients}
\usage{
locateMTvariants(query)
}
\arguments{
\item{query}{A single variant in MVRanges form}
}
\value{
An MVRanges with start and end codons (overlaps return 2 copies of the variant)
}
\description{
Locates the variant and determines local bounds within gene
Primarily meant to be used within injectMTvarients
}
