% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{MAlignments-methods}
\alias{MAlignments-methods}
\alias{coverage,MAlignments-method}
\alias{runLength,MAlignments-method}
\alias{runValue,MAlignments-method}
\alias{Summary,MAlignments-method}
\alias{show,MAlignments-method}
\alias{fileName,MAlignments-method}
\alias{scanBamHeader,MAlignments-method}
\title{MAlignments methods (centralized).}
\usage{
\S4method{coverage}{MAlignments}(x)

\S4method{runLength}{MAlignments}(x)

\S4method{runValue}{MAlignments}(x)

\S4method{Summary}{MAlignments}(x)

\S4method{show}{MAlignments}(object)

\S4method{fileName}{MAlignments}(object)

\S4method{scanBamHeader}{MAlignments}(files)
}
\arguments{
\item{x}{an MAlignments}

\item{object}{an MAlignments}

\item{files}{an MAlignments}
}
\value{
various things, as appropriate to the methods
}
\description{
Depending on how a generic was originally designated, the arguments to 
these methods can have various argument names, but all of them tend to 
take an MAlignments as their argument.
}
