% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\name{MAlignments}
\alias{MAlignments}
\title{wrap a GAlignments for easier stats}
\usage{
MAlignments(gal, bam)
}
\arguments{
\item{gal}{a GAlignments}

\item{bam}{a bam filename}
}
\value{
an MAlignments
}
\description{
wrap a GAlignments for easier stats
}
