% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{class}
\name{MAlignmentsList-class}
\alias{MAlignmentsList-class}
\title{wraps a GAlignmentsList (made up of MAlignments) for nicer viewing}
\description{
This is largely deprecated with the move from gmapR to Rsamtools::pileup.
}
