% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{MAlignmentsList-methods}
\alias{MAlignmentsList-methods}
\alias{coverage,MAlignmentsList-method}
\alias{runLength,MAlignmentsList-method}
\alias{fileName,MAlignmentsList-method}
\alias{Summary,MAlignmentsList-method}
\alias{show,MAlignmentsList-method}
\title{MAlignmentsList methods (centralized).}
\usage{
\S4method{coverage}{MAlignmentsList}(x)

\S4method{runLength}{MAlignmentsList}(x)

\S4method{fileName}{MAlignmentsList}(object)

\S4method{Summary}{MAlignmentsList}(x)

\S4method{show}{MAlignmentsList}(object)
}
\arguments{
\item{x}{an MAlignmentsList}

\item{object}{an MAlignmentsList}
}
\value{
various objects, as appropriate to the method
}
\description{
Depending on how a generic was originally designated, the arguments to 
these methods can have various argument names, but all of them tend to 
take an MAlignmentsList as their argument.
}
