% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\name{MAlignmentsList}
\alias{MAlignmentsList}
\title{wrap a GAlignmentsList for viewing}
\usage{
MAlignmentsList(...)
}
\arguments{
\item{...}{MAlignments}
}
\value{
an MAlignments
}
\description{
wrap a GAlignmentsList for viewing
}
