% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtCircos.R
\name{mtCircos}
\alias{mtCircos}
\title{plot a canonical human (or, in principle, any) mitochondrial genome}
\usage{
mtCircos(anno = NULL, variants = NULL, matrices = NULL)
}
\arguments{
\item{anno}{a GRanges (optional, defaults to mtAnno.rCRS if none given)}

\item{variants}{optional MVRanges or MVRangesList to plot outside the circle}

\item{matrices}{optional matrices of data to plot inside of the genes circle}
}
\value{
invisibly, a list: `anno` (data.frame) + `pfun` (panel.fun)
}
\description{
The default font sizes, orientations, etc. are optimized for a "cold" start;
if you want to fiddle with the details, crack open the code and modify it...
or alternatively, add sectors/dendrograms inside of this "framed" version.
}
