% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragePlot.R
\name{mtCoveragePlot}
\alias{mtCoveragePlot}
\title{plot read coverage across the mitochondrial genome}
\usage{
mtCoveragePlot(mal, bases = NULL)
}
\arguments{
\item{mal}{an MAlignments (NOT A LIST)}
}
\description{
plot read coverage across the mitochondrial genome
}
