% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtGenes.rCRS.R
\docType{data}
\name{mtGenes.rCRS}
\alias{mtGenes.rCRS}
\title{All annotated gene regions (not just coding genes) from rCRS.}
\format{a GRanges}
\usage{
mtGenes.rCRS
}
\description{
All annotated gene regions (not just coding genes) from rCRS.
}
\examples{
data(mtGenes.rCRS) 
subset(mtGenes.rCRS, region == "coding")
subset(mtGenes.rCRS, region == "tRNA") 

}
\keyword{datasets}
