% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtGenes.R
\docType{data}
\name{mtGenes}
\alias{mtGenes}
\title{Base sequences of mitochondrial coding genes.}
\format{
a GRanges with a DNAStringSet of the base sequences as its mcols.
}
\usage{
mtGenes
}
\description{
Base sequences of mitochondrial coding genes.
}
\examples{
data(mtGenes)
width(mtGenes)
names(mtGenes$DNA)

}
\keyword{datasets}
