% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtHGVS.R
\name{mtHGVS}
\alias{mtHGVS}
\title{convert mitochondrial variant calls to HGVS format for naming}
\usage{
mtHGVS(x)
}
\arguments{
\item{x}{an MVRanges (or, in a pinch, a GRanges)}
}
\value{
a properly named MVRanges (or, in a pinch, GRanges)
}
\description{
convert mitochondrial variant calls to HGVS format for naming
}
