% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRanges.R
\name{MVRanges}
\alias{MVRanges}
\title{wrap a VRanges for mitochondrial use}
\usage{
MVRanges(vr, coverage)
}
\arguments{
\item{vr}{the VRanges}

\item{covg}{estimated coverage}
}
\value{
an MVRanges
}
\description{
wrap a VRanges for mitochondrial use
}
