% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{MVRangesList-methods}
\alias{MVRangesList-methods}
\alias{coverage,MVRangesList-method}
\alias{genes,MVRangesList-method}
\alias{snpCall,MVRangesList-method}
\alias{getAnnotations,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{predictCoding,MVRangesList,missing,missing,missing-method}
\alias{show,MVRangesList-method}
\alias{filt,MVRangesList-method}
\alias{granges,MVRangesList-method}
\alias{summarizeVariants,MVRangesList,missing,missing-method}
\alias{genome,MVRangesList-method}
\alias{locateVariants,MVRangesList,missing,missing-method}
\alias{tallyVariants,MVRangesList-method}
\alias{plot,MVRangesList,ANY-method}
\title{MVRangesList methods (centralized).}
\usage{
\S4method{coverage}{MVRangesList}(x)

\S4method{genes}{MVRangesList}(x)

\S4method{snpCall}{MVRangesList}(object)

\S4method{getAnnotations}{MVRangesList}(annotations)

\S4method{encoding}{MVRangesList}(x)

\S4method{predictCoding}{MVRangesList,missing,missing,missing}(query,
  subject, seqSource, varAllele, ...)

\S4method{show}{MVRangesList}(object)

\S4method{filt}{MVRangesList}(x)

\S4method{granges}{MVRangesList}(x, filterLowQual = TRUE)

\S4method{summarizeVariants}{MVRangesList,missing,missing}(query,
  filterLowQual = TRUE, ...)

\S4method{genome}{MVRangesList}(x)

\S4method{locateVariants}{MVRangesList,missing,missing}(query,
  filterLowQual = TRUE, ...)

\S4method{tallyVariants}{MVRangesList}(x, filterLowQual = TRUE, ...)

\S4method{plot}{MVRangesList,ANY}(x, y, ...)
}
\arguments{
\item{x}{an MVRangesList (for some methods)}

\item{object}{an MVRangesList (for other methods)}

\item{annotations}{an MVRangesList (for getAnnotations)}

\item{query}{an MVRangesList (for predictCoding)}

\item{subject}{a GRanges, usually}

\item{seqSource}{a BSgenome, usually}

\item{varAllele}{variant alleles}

\item{...}{miscellaneous args, passed through}

\item{filterLowQual}{opt. for `granges`/`summarizeVariants`/`tallyVariants`}

\item{y}{another MVRangesList}
}
\value{
depends on the method invoked.
}
\description{
MVRangesList methods (centralized).
}
\section{Utility methods}{


`coverage`             returns estimated mitochondrial read coverage depth
`filt`                 removes variants where PASS != TRUE for each element
}

\section{Annotation methods}{


`genes`                returns an annotated GRanges of mitochondrial genes 
`getAnnotations`       returns a GRanges of annotated mitochondrial features
`genome`               returns the genome (or, perhaps, genomes) in an MVRL
`encoding`             returns mutations in coding regions for each element
`granges`              returns mildly annotated aggregates of variant sites
`snpCall`              retrieves single nucleotide variant polymorphisms 
`tallyVariants`        return a matrix of variant types by annotated region
`locateVariants`       locates variants within genes, tRNA, rRNA, or D-loop
`summarizeVariants`    attempts mass functional annotation of variant sites
}

\section{Visualization methods}{


`plot`                 creates circular plot of mitochondrial variant calls
}

