% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawMTreads.R
\name{rawMTreads}
\alias{rawMTreads}
\title{Raw reads shown as GAlignments\link{List}}
\usage{
rawMTreads(bams)
}
\arguments{
\item{bams}{List of files}
}
\value{
\preformatted{   GAlignments or GAlignmentsList of raw reads
}
}
\description{
Raw reads shown as GAlignments\link{List}
}
\examples{

library(MTseekerData)
BAMdir <- system.file("extdata", "BAMs", package="MTseekerData")
BAMs <- list.files(BAMdir, pattern="bam$")
BAMs <- paste0(BAMdir, "/", BAMs)
rawReads <- rawMTreads(BAMs)
rawReads_1 <- MAlignments(rawReads[[1]], BAMs[1])
plotMTCoverage(rawReads_1)

}
