% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCRSeq.R
\docType{data}
\name{rCRSeq}
\alias{rCRSeq}
\title{The complete sequence of the human rCRS mitogenome. Yes, it's that small.}
\format{a DNAStringSet of length 1}
\usage{
rCRSeq
}
\description{
The complete sequence of the human rCRS mitogenome. Yes, it's that small.
}
\examples{
data(rCRSeq) 
width(rCRSeq) 
data(mtGenes.rCRS) 
getSeq(rCRSeq, mtGenes.rCRS)

}
\keyword{datasets}
