% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4Methods.R
\name{s4Methods}
\alias{s4Methods}
\title{List all defined methods for an S4 class (or classes, if you must)}
\usage{
s4Methods(...)
}
\arguments{
\item{...}{name[s] of class[es] (please note, results will be union'ed)}
}
\value{
methods for the class[es], union'ed into a character vector
}
\description{
Mostly for debugging and making architectural choices, e.g. about coverage()
}
\details{
Note: this is borrowed from Hadley, who borrowed it from a BioC workshop!
}
\examples{
library(MTseeker) 
s4methods("MVRangesList")
s4methods("MAlignmentsList")

}
