% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validMetadata.R
\name{validMetadata}
\alias{validMetadata}
\title{Ensure that the metadata caches in MAlignmentsLists and MVRangesLists are OK}
\usage{
validMetadata(x)
}
\arguments{
\item{x}{an MAlignmentsList or an MVRangesList}
}
\value{
\preformatted{       TRUE or FALSE (if FALSE, attr(res)$mismatches shows why)
}
}
\description{
In order to avoid a lot of lengthy calculations, both MAlignmentsList and
MVRangesList objects keep a cache of some relevant statistics and filenames
in their metadata slot. If these caches get stale, it can cause problems.
}
\details{
This function performs some sanity checks on the caches so that the above
problems are unlikely to occur, provided that checkMetadataCache() is called
at sensible times. This function is NOT a replacement for validObject().
}
\examples{
library(MTseekerData)
data(RONKSvariants) 
obj <- "RONKSvariants"
if (validMetadata(get(obj))) message(obj, " has valid metadata.")
}
