% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVCF.R
\name{vcf2mvrl}
\alias{vcf2mvrl}
\title{Convert a VCF to an MVRangesList object}
\usage{
vcf2mvrl(
  vcf_filename,
  bam_filename = NULL,
  gzip = TRUE,
  biscuit = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{vcf_filename}{A filename/path to the corresponding VCF}

\item{bam_filename}{A filename/path to the corresponding BAM}

\item{gzip}{Is the input gzipped? (TRUE; really ought not matter)}

\item{biscuit}{Is the input VCF from biscuit? (FALSE)}

\item{verbose}{Print all messages during progress? (TRUE)}
}
\value{
an MVRangesList
}
\description{
Convert a VCF to an MVRangesList object
}
\examples{

library(MTseekerData)
VCFdir <- system.file("extdata", package="MTseekerData")
VCF <- file.path(VCFdir, list.files(VCFdir, pattern="*.vcf.gz$"))
mvr <- vcf2mvrl(VCF)[[1]]
mvr$FILTER <- ifelse(totalDepth(mvr) >= 20, "PASS", ".")
mvr$PASS <- mvr$FILTER == "PASS"
filterMT(mvr)

}
