% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byStrand.R
\name{byStrand}
\alias{byStrand}
\title{simple helper function to split a *RangesList by the strand of its mt target}
\usage{
byStrand(x, anno = NULL)
}
\arguments{
\item{x}{a *RangesList}

\item{anno}{optional feature annotation, will use mtAnno.rCRS if NULL}
}
\value{
a list with elements of x overlapping features on each strand
}
\description{
simple helper function to split a *RangesList by the strand of its mt target
}
