% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMT.R
\name{callMT}
\alias{callMT}
\alias{callMTVars}
\title{call mitochondrial variants against rCRS from an MAlignments[List] object}
\usage{
callMT(mal, ..., parallel = FALSE, verbose = FALSE)

callMTVars(BAM, SIZE = 75, GENOME = "rCRS", CHR = "chrM",
  COV = NULL, verbose = FALSE)
}
\arguments{
\item{mal}{an MAlignments (or, potentially, an MAlignmentsList)}

\item{...}{other arguments to pass to VariantTools::callVariants}

\item{parallel}{try to run in parallel? (FALSE; this is super unstable)}

\item{verbose}{be verbose? (FALSE; turn on for debugging purposes)}

\item{BAM}{the BAM filename (for callMTVars)}

\item{SIZE}{the read length (for callMTVars; default is 75)}

\item{GENOME}{the reference genome (for callMTVars; default is rCRS)}

\item{CHR}{the mt contig name (for callMTVars; default is chrM)}

\item{COV}{average read coverage (so we don't have to countBam)}
}
\value{
an MVRanges (or, potentially, an MVRangesList)
}
\description{
`callMTVars` is a helper function for callMT
}
\details{
FIXME: transition gmapR from import to suggestion; use tallyVariants alone?

FIXME: use Rsamtools::pileup by default and only use gmapR if requested

FIXME: figure out how to directly call variants off [GM]Alignments[List]s

FIXME: figure out a way to reprocess extracted chrM/MT reads against rCRS,
       regardless of what reference they were originally aligned against.
}
\examples{
library(MTseekerData)
data(RONKSreads)
\donttest{
  # requires the BAMs
  # also currently requires gmapR 
  RONKSvariants <- callMT(RONKSreads) 
}
head(RONKSvariants$RO_1) 

}
