% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMT.R
\name{filterMT}
\alias{filterMT}
\title{Filter a RangedSummarizedExperiment's columns based on its $mtCovg colData}
\usage{
filterMT(SE, minCovg = 20)
}
\arguments{
\item{SE}{a RangedSummarizedExperiment with colData() named mtCovg}

\item{minCovg}{minimum covg (20, cf. Griffin, Genetics in Medicine 2014)}
}
\description{
Griffin et al. (Genetics in Medicine 2014) recommends 20x coverage for mtDNA
sequencing to have comparable error rates to Sanger sequencing.  By default, 
that is the cutoff applied here to ensure halfway decent variant annotation.
}
