% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMTvars.R
\name{filterMTvars}
\alias{filterMTvars}
\title{sanitize PASSing mitochondrial variant calls to a moderate degree}
\usage{
filterMTvars(vars, fp = TRUE, NuMT = 0.03, covg = 20)
}
\arguments{
\item{vars}{an MVRanges or MVRangesList (will be unlisted and relisted)}

\item{fp}{standard issue false positive filters? (TRUE: use Triska & RSRS)}

\item{NuMT}{variants with VAF < this number will be presumed NuMTs (0.03)}

\item{covg}{minimum median read coverage across chrM to be considered (20)}
}
\value{
a filtered set of variants
}
\description{
sanitize PASSing mitochondrial variant calls to a moderate degree
}
\examples{
library(MTseekerData) 
filterMTvars(RONKSvariants$RO_1)

}
