% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMT.R
\name{getMT}
\alias{getMT}
\title{grab the mitochondrial reads from a BAM & estimate their fraction (of total)}
\usage{
getMT(bam, filter = FALSE, plotMAPQ = FALSE, ...)
}
\arguments{
\item{bam}{a BAM filename, or DataFrame/SummarizedExperiment with $BAM}

\item{filter}{filter on bam$mtCovg? (default is FALSE, don't filter)}

\item{plotMAPQ}{plot distribution of mitochondrial mapping quality? (FALSE)}

\item{...}{additional args to pass scanBamParam(), such as mapqFilter}
}
\value{
\preformatted{     an MAlignments or MAlignmentsList object
}
}
\description{
This purely a convenience function, and an incredibly convenient one at that.
FIXME: this function is still useful, but it desperately needs refactoring.
}
\examples{
library(MTseekerData)
BAMdir <- system.file("extdata", "BAMs", package="MTseekerData")
patientBAMs <- paste0(BAMdir, "/", list.files(BAMdir, pattern="^pt.*bam$"))
(mal <- getMT(patientBAMs[1]))
class(mal) 
}
