% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploMask.R
\name{haploMask}
\alias{haploMask}
\title{Mask haplogroup specific variants using haplogrep}
\usage{
haploMask(mvr, fasta.output = NULL, mask = TRUE,
  return.haplogroup = TRUE, override = FALSE)
}
\arguments{
\item{mvr}{Input should be either an MVRanges or MVRangesList}

\item{fasta.output}{Path to a location to output the per-sample consensus variant fasta files}

\item{mask}{Whether to mask out the haplogroup specific SNPs}

\item{return.haplogroup}{Return the inferred haplogroup}

\item{override}{This will override the java check and proceed even if a java install isn't detected}
}
\value{
An MVRanges or MVRangesList of masked haplogroup-specific variants
}
\description{
Mask haplogroup specific variants using haplogrep
}
\examples{

}
