% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectMTvariants.R
\name{injectMTVariants}
\alias{injectMTVariants}
\title{Inject (one or more) variants against rCRS.}
\usage{
injectMTVariants(mvr, coding = TRUE, gr = NULL)
}
\arguments{
\item{mvr}{An MVRanges, usually from pileupMT, often subsetted}

\item{coding}{TRUE to only look at coding regions}

\item{gr}{A GRanges, usually of protein-coding regions (the default)}
}
\value{
\preformatted{     The GRanges, with ref/var DNA and AA and 
}
}
\description{
FIXME: this function could most likely be orders of magnitude faster.
FIXME: this ONLY considers variants injected against rCRS, not RSRS or hg19.
FIXME: this function really needs mouse support as well.
}
\examples{
library(MTseekerData)
injectMTVariants(RONKSvariants[["RO_2"]]) 

}
