% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injectMtVariants.R
\name{injectMtVariants}
\alias{injectMtVariants}
\title{Inject (one or more) variants against rCRS.}
\usage{
injectMtVariants(gr, mvr, xlate = TRUE)
}
\arguments{
\item{gr}{A GRanges, usually of protein-coding regions}

\item{mvr}{An MVRanges, usually from callMT}

\item{xlate}{Attempt to translate codon(s) affected by variant(s)? (TRUE)}
}
\value{
The same GRanges, but with ref/var DNA and AA from rCRS.
}
\description{
FIXME: this function could most likely be orders of magnitude faster.
FIXME: this ONLY considers variants injected against rCRS, not RSRS or hg19.
}
