% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locateMTvariants.R
\name{locateMTvariants}
\alias{locateMTvariants}
\title{Locates the variant and determines local bounds within gene
Primarily meant to be used within injectMTvarients}
\usage{
locateMTvariants(query, coding = TRUE)
}
\arguments{
\item{query}{A single variant in MVRanges form}

\item{coding}{Look only at coding regions? (TRUE)}
}
\value{
\preformatted{      An MVRanges with start/end codons (overlaps return copies)
}
}
\description{
Locates the variant and determines local bounds within gene
Primarily meant to be used within injectMTvarients
}
