% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{class}
\name{MAlignments-class}
\alias{MAlignments-class}
\title{wraps a GAlignments with information about coverage and its target BAM file}
\description{
wraps a GAlignments with information about coverage and its target BAM file
}
