% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignments.R
\docType{methods}
\name{MAlignments-methods}
\alias{MAlignments-methods}
\alias{coverage,MAlignments-method}
\alias{runLength,MAlignments-method}
\alias{runValue,MAlignments-method}
\alias{Summary,MAlignments-method}
\alias{show,MAlignments-method}
\alias{fileName,MAlignments-method}
\alias{scanBamHeader,MAlignments-method}
\title{MAlignments methods (centralized).}
\usage{
\S4method{coverage}{MAlignments}(x)

\S4method{runLength}{MAlignments}(x)

\S4method{runValue}{MAlignments}(x)

\S4method{Summary}{MAlignments}(x)

\S4method{show}{MAlignments}(object)

\S4method{fileName}{MAlignments}(object)

\S4method{scanBamHeader}{MAlignments}(files)
}
\arguments{
\item{x}{an MAlignments}

\item{x}{an MAlignments}

\item{x}{an MAlignments}

\item{x}{an MAlignments}

\item{x}{an MAlignments}

\item{file}{an MAlignments}

\item{bamRanges}{an MAlignments}
}
\value{
estimated coverage (numeric)

read length

genome size

summary of x

a BAM filename (string)

a BAM header
}
\description{
MAlignments methods (centralized).
}
