% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\name{MAlignmentsList}
\alias{MAlignmentsList}
\alias{MAlignmentsList-methods}
\alias{show,MAlignmentsList-method}
\alias{fileName,MAlignmentsList-method}
\alias{scanBamHeader,MAlignmentsList-method}
\alias{readLength,MAlignmentsList-method}
\alias{genomeLength,MAlignmentsList-method}
\alias{genomeCoverage,MAlignmentsList-method}
\title{wrap a GAlignmentsList for viewing}
\usage{
MAlignmentsList(...)

\S4method{show}{MAlignmentsList}(object)

\S4method{fileName}{MAlignmentsList}(object)

\S4method{scanBamHeader}{MAlignmentsList}(files)

\S4method{readLength}{MAlignmentsList}(x)

\S4method{genomeLength}{MAlignmentsList}(x)

\S4method{genomeCoverage}{MAlignmentsList}(x)
}
\arguments{
\item{...}{MAlignments}

\item{object}{an MAlignmentsList}

\item{files}{an MAlignmentsList}

\item{x}{an MAlignmentsList}
}
\value{
\preformatted{       an MAlignments 
}

\preformatted{     various objects, as appropriate to the method 
}
}
\description{
Normally the MAlignmentsList constructor will be called by getMT.

Depending on how a generic was originally designated, the arguments to
these methods can have various argument names, but all of them tend to
take an MAlignmentsList as their argument.
}
\examples{

library(MTseekerData)
BAMdir <- system.file("extdata", "BAMs", package="MTseekerData")
print(BAMdir)
BAMs <- paste0(BAMdir, "/", list.files(BAMdir, pattern=".bam$"))
print(BAMs)
mall <- getMT(BAMs[1]) # a xenograft
genomeLength(mall) 
readLength(mall) 
fileName(mall) 
scanBamHeader(mall) 

}
