% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{MAlignmentsList}
\alias{MAlignmentsList}
\alias{MAlignmentsList-methods}
\alias{genomeCoverage,MAlignmentsList-method}
\alias{readLength,MAlignmentsList-method}
\alias{fileName,MAlignmentsList-method}
\alias{Summary,MAlignmentsList-method}
\alias{show,MAlignmentsList-method}
\title{wrap a GAlignmentsList for viewing}
\usage{
MAlignmentsList(...)

\S4method{genomeCoverage}{MAlignmentsList}(x)

\S4method{readLength}{MAlignmentsList}(x)

\S4method{fileName}{MAlignmentsList}(object)

\S4method{Summary}{MAlignmentsList}(x)

\S4method{show}{MAlignmentsList}(object)
}
\arguments{
\item{...}{MAlignments}

\item{x}{an MAlignmentsList}

\item{object}{an MAlignmentsList}
}
\value{
\preformatted{       an MAlignments 
}

\preformatted{     various objects, as appropriate to the method 
}
}
\description{
Normally the MAlignmentsList constructor will be called by getMT.

Depending on how a generic was originally designated, the arguments to
these methods can have various argument names, but all of them tend to
take an MAlignmentsList as their argument.
}
\examples{
\dontrun{
library(MTseekerData)
BAMdir <- system.file("extdata", "BAMs", package="MTseekerData")
print(BAMdir)
BAMs <- paste0(BAMdir, "/", list.files(BAMdir, pattern=".bam$"))
print(BAMs)
targets <- data.frame(BAM=BAMs, stringsAsFactors=FALSE) 
rownames(targets) <- sapply(strsplit(basename(BAMs), "\\\\."), `[`, 1)
mall <- getMT(targets)
class(mall) 
show(mall) 
}
}
