% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAlignmentsList.R
\docType{methods}
\name{MAlignmentsList-methods}
\alias{MAlignmentsList-methods}
\alias{updateObject,MAlignmentsList-method}
\alias{coverage,MAlignmentsList-method}
\alias{runLength,MAlignmentsList-method}
\alias{fileName,MAlignmentsList-method}
\alias{Summary,MAlignmentsList-method}
\alias{show,MAlignmentsList-method}
\title{MAlignmentsList methods (centralized).}
\usage{
\S4method{updateObject}{MAlignmentsList}(object)

\S4method{coverage}{MAlignmentsList}(x)

\S4method{runLength}{MAlignmentsList}(x)

\S4method{fileName}{MAlignmentsList}(object)

\S4method{Summary}{MAlignmentsList}(x)

\S4method{show}{MAlignmentsList}(object)
}
\arguments{
\item{object}{an MAlignmentsList, usually lacking some cached information}

\item{x}{an MAlignmentsList}

\item{x}{an MAlignmentsList}

\item{object}{an MAlignmentsList}

\item{x}{an MAlignmentsList}

\item{objects}{an MAlignmentsList}
}
\value{
an updated MAlignmentsList with cached metadata summaries

estimated coverage (numeric vector)

estimated coverage (numeric vector)

BAM file summary for the MAlignmentsList

a DataFrame
}
\description{
MAlignmentsList methods (centralized).
}
