% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtCircos.R
\name{mtCircos}
\alias{mtCircos}
\title{plot a canonical human (or, in principle, any) mitochondrial genome}
\usage{
mtCircos(anno = NULL, variants = NULL)
}
\arguments{
\item{anno}{a GRanges (optional, defaults to mtAnno.rCRS if none given)}

\item{variants}{an MVRanges or MVRangesList (optional, and poorly supported)}
}
\value{
a list with elements `anno` (data.frame) and `pfun` (panel.fun)
}
\description{
The default font sizes, orientations, etc. are optimized for a "cold" start;
if you want to fiddle with the details, crack open the code and modify it...
or alternatively, add sectors/dendrograms inside of this "framed" version.
}
