% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtCircos.R
\name{mtCircos}
\alias{mtCircos}
\title{plot a canonical human (or, in principle, any) mitochondrial genome}
\usage{
mtCircos(variants = NULL, outside = NULL, inside = NULL, outcol = NULL,
  incol = NULL, anno = NULL, how = c("matrix", "VAF"), ...)
}
\arguments{
\item{variants}{optional MVRanges or MVRangesList to split by strand & plot}

\item{outside}{optional MVRanges or MVRangesList to plot outside the circle}

\item{inside}{optional MVRanges or MVRangesList to plot inside the circle}

\item{outcol}{optional color assignment function or matrix for outside}

\item{incol}{optional color assignment function or matrix for inside}

\item{anno}{a GRanges (optional, defaults to mtAnno.rCRS if none given)}

\item{how}{optional specification for how to plot multiple samples}

\item{...}{other arguments to pass on to called functions}
}
\value{
invisibly, a list: `anno` (data.frame) + `pfun` (panel.fun)
}
\description{
The default font sizes, orientations, etc. are optimized for a "cold" start;
if you want to fiddle with the details, crack open the code and modify it...
or alternatively, add sectors/dendrograms inside of this "framed" version.
}
\details{
FIXME: add variant type coloration (del=blue, SNV=black, ins=red)
}
