% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTcomplex.R
\name{MTcomplex}
\alias{MTcomplex}
\title{plot the (putative) functional impact of mutations to ETP genes as SVG}
\usage{
MTcomplex(variants, defColor = "#c9eded", verbose = FALSE)
}
\arguments{
\item{variants}{an MVRanges or MVRangesList}

\item{defColor}{default color (#c9eded is standard)}

\item{verbose}{be verbose, for debugging? (FALSE)}
}
\value{
\preformatted{     invisibly, the temporary file to which the SVG was written
}
}
\description{
Tim Vickers created a beautiful illustration of the mitochondrial electron
transport chain, and that's where coding mitochondrial DNA mutations will
usually hit (we aren't plotting the mitoribosome or tRNAs just yet). So why
reinvent the wheel (and possibly make it square)?
}
\examples{
library(MTseekerData)
data(RONKSvariants) 
MTcomplex(RONKSvariants$RO_1)

}
