% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtCoverage.R
\name{mtCoverage}
\alias{mtCoverage}
\alias{plotMtCoverage}
\title{mitochondrial genome coverage for an MAlignments or MVRanges object}
\usage{
mtCoverage(x, ...)

plotMtCoverage(x, ...)
}
\arguments{
\item{x}{an MAlignments or MVRanges}

\item{...}{other arguments to pass to GenomicAlignments::coverage()}
}
\value{
an RleList
}
\description{
We co-opted the `coverage` method to retrieve approximate coverage depth
across the mitochondrial genome in MAlignments[List] and MVRanges[list],
so this function gives back what it was supposed to do (provide an Rle) 
and can allow for some subsetting (e.g. variant-supporting-read coverage)
that may be of interest when interpreting results.
}
\details{
plotMtCoverage does what one might expect, and plots read coverage.
}
\examples{
library(MTseekerData)

data(RONKSreads)
mtCoverage(RONKSreads$RO_1)
plotMtCoverage(RONKSreads$RO_1)

data(RONKSvariants)
mtCoverage(RONKSvariants$RO_1)
plotMtCoverage(RONKSvariants$RO_1)

par(mfrow=c(1,2))
plotMtCoverage(RONKSreads$NKs_1)
title("Read coverage for normal kidney sample 1") 
plotMtCoverage(RONKSreads$RO_1)
title("Read coverage for renal oncocytoma sample 1") 

}
