% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTcoverage.R
\name{MTcoverage}
\alias{MTcoverage}
\alias{plotMTCoverage}
\alias{plotStrandedMTCoverage}
\title{Mitochondrial genome coverage and plots for MAlignments or MVRanges objects}
\usage{
MTcoverage(x, ...)

plotMTCoverage(x, ref = c("rCRS", "NC_005089"), ...)

plotStrandedMTCoverage(x, ...)
}
\arguments{
\item{x}{an MAlignments or MVRanges}

\item{...}{other arguments to pass to GenomicAlignments::coverage()}

\item{ref}{string denoting reference sequence to use}
}
\value{
\preformatted{     an RleList (or, invisibly for plot functions, a result list)
}
}
\description{
We co-opted the \code{coverage} method to retrieve approximate coverage depth
across the mitochondrial genome in MAlignments\link{List} and MVRanges\link[=List]{list},
so this function gives back what it was supposed to do (provide an Rle)
and can allow for some subsetting (e.g. variant-supporting-read coverage)
that may be of interest when interpreting results.
}
\details{
The plotting functions can handle MAlignments or MVRanges objects directly.

plotMTCoverage does what one might expect, and plots (read or call) coverage.

plotStrandedMTCoverage does the same thing, but keeps track of which strand.
}
\examples{
library(MTseekerData)

data(RONKSreads)
MTcoverage(RONKSreads$RO_1)
plotMTCoverage(RONKSreads$RO_1)

data(RONKSvariants)
MTcoverage(RONKSvariants$RO_1)
plotMTCoverage(RONKSvariants$RO_1)

par(mfrow=c(1,2))
plotMTCoverage(RONKSreads$NKS_1)
title("Read coverage for normal kidney sample 1") 
plotMTCoverage(RONKSreads$RO_1)
title("Read coverage for renal oncocytoma sample 1") 

par(mfrow=c(1,2))
plotStrandedMTCoverage(RONKSreads$NKS_1)
title("Stranded read coverage for normal kidney sample 1") 
plotStrandedMTCoverage(RONKSreads$RO_1)
title("Stranded read coverage for renal oncocytoma sample 1") 

}
