% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtHGVS.R
\name{mtHGVS}
\alias{mtHGVS}
\title{convert mitochondrial variant calls to HGVS format for naming}
\usage{
mtHGVS(x, asMVR = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{an MVRanges (or, in a pinch, a GRanges)}

\item{asMVR}{return a renamed MVRanges? (default is FALSE)}

\item{verbose}{be yappy? (default is FALSE)}
}
\value{
proper HGVS names for the *Ranges (or a renamed *Ranges)
}
\description{
convert mitochondrial variant calls to HGVS format for naming
}
\examples{
library(MTseekerData) 
data(RONKSvariants) 
mtHGVS(RONKSvariants$RO_1)

}
