% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtHGVS.R
\name{mtHGVS}
\alias{mtHGVS}
\title{convert mitochondrial variant calls to HGVS format for naming}
\usage{
mtHGVS(x, verbose = FALSE)
}
\arguments{
\item{x}{an MVRanges (or, in a pinch, a GRanges)}

\item{verbose}{be yappy? (default is FALSE)}
}
\value{
proper HGVS names for the MVRanges (or GRanges)
}
\description{
convert mitochondrial variant calls to HGVS format for naming
}
