% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtHGVS.R
\name{mtHGVS}
\alias{mtHGVS}
\title{convert mitochondrial variant calls to HGVS format for naming}
\usage{
mtHGVS(mvr)
}
\arguments{
\item{mvr}{an MVRanges}
}
\value{
a properly named MVRanges
}
\description{
convert mitochondrial variant calls to HGVS format for naming
}
