% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\docType{methods}
\name{MVRangesList-methods}
\alias{MVRangesList-methods}
\alias{coverage,MVRangesList-method}
\alias{counts<-,MVRangesList,RangedSummarizedExperiment-method}
\alias{counts,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{predictCoding,MVRangesList,missing,missing,missing-method}
\alias{show,MVRangesList-method}
\alias{filt,MVRangesList-method}
\alias{granges,MVRangesList-method}
\alias{summarizeVariants,MVRangesList,missing,missing-method}
\alias{genome,MVRangesList-method}
\alias{locateVariants,MVRangesList,missing,missing-method}
\alias{tallyVariants,MVRangesList-method}
\title{MVRangesList methods (centralized).}
\usage{
\S4method{coverage}{MVRangesList}(x)

\S4method{counts}{MVRangesList,RangedSummarizedExperiment}(object) <- value

\S4method{counts}{MVRangesList}(object)

\S4method{encoding}{MVRangesList}(x)

\S4method{predictCoding}{MVRangesList,missing,missing,missing}(query, subject,
  seqSource, varAllele, ...)

\S4method{show}{MVRangesList}(object)

\S4method{filt}{MVRangesList}(x)

\S4method{granges}{MVRangesList}(x, filterLowQual = TRUE)

\S4method{summarizeVariants}{MVRangesList,missing,missing}(query,
  filterLowQual = TRUE, ...)

\S4method{genome}{MVRangesList}(x)

\S4method{locateVariants}{MVRangesList,missing,missing}(query,
  filterLowQual = TRUE, ...)

\S4method{tallyVariants}{MVRangesList}(x, filterLowQual = TRUE, ...)
}
\arguments{
\item{x}{an MVRangesList (for some methods)}

\item{object}{an MVRangesList (for other methods)}

\item{value}{a RangedSummarizedExperiment with matching colnames}

\item{query}{an MVRangesList (for predictCoding)}

\item{filterLowQual}{opt. for `granges`/`summarizeVariants`/`tallyVariants`}
}
\description{
`counts`               returns fragment counts, if any
`counts<-`             adds or updates fragment counts
`coverage`             returns estimated coverage for each element
`genome`               returns the genome (or, perhaps, genomes) in an MVRL
`filt`                 removes variants where PASS != TRUE for each element 
`encoding`             returns mutations in coding regions for each element
`granges`              returns mildly annotated aggregates of variant sites
`tallyVariants`        return a matrix of variant types by annotated region
`locateVariants`       locates variants within genes, tRNA, rRNA, or D-loop
`summarizeVariants`    attempts mass functional annotation of variant sites
}
