% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVRangesList.R
\name{MVRangesList}
\alias{MVRangesList}
\alias{MVRangesList-methods}
\alias{genomeCoverage,MVRangesList-method}
\alias{genes,MVRangesList-method}
\alias{snpCall,MVRangesList-method}
\alias{getAnnotations,MVRangesList-method}
\alias{encoding,MVRangesList-method}
\alias{coverage,MVRangesList-method}
\alias{predictCoding,MVRangesList,missing,missing,missing-method}
\alias{show,MVRangesList-method}
\alias{filt,MVRangesList-method}
\alias{granges,MVRangesList-method}
\alias{summarizeVariants,MVRangesList,missing,missing-method}
\alias{genome,MVRangesList-method}
\alias{locateVariants,MVRangesList,missing,missing-method}
\alias{plot,MVRangesList,ANY-method}
\alias{consensusString,MVRangesList-method}
\title{Wrap a VRangesList for mitochondrial use.}
\usage{
MVRangesList(..., bamFiles = NULL, coverageRles = NULL, verbose = FALSE)

\S4method{genomeCoverage}{MVRangesList}(x)

\S4method{genes}{MVRangesList}(x)

\S4method{snpCall}{MVRangesList}(object)

\S4method{getAnnotations}{MVRangesList}(annotations)

\S4method{encoding}{MVRangesList}(x)

\S4method{coverage}{MVRangesList}(x)

\S4method{predictCoding}{MVRangesList,missing,missing,missing}(query, subject, seqSource, varAllele, ...)

\S4method{show}{MVRangesList}(object)

\S4method{filt}{MVRangesList}(x)

\S4method{granges}{MVRangesList}(x, filterLowQual = TRUE)

\S4method{summarizeVariants}{MVRangesList,missing,missing}(query, filterLowQual = TRUE, ...)

\S4method{genome}{MVRangesList}(x)

\S4method{locateVariants}{MVRangesList,missing,missing}(query, filterLowQual = TRUE, ...)

\S4method{plot}{MVRangesList,ANY}(x, y, ...)

\S4method{consensusString}{MVRangesList}(x, ...)
}
\arguments{
\item{...}{miscellaneous args, passed through}

\item{bamFiles}{(Optional) BAM filenames for each MVRanges element}

\item{coverageRles}{(Optional) coverage Rles for each MVRanges element}

\item{verbose}{Be verbose? (default is FALSE)}

\item{x}{an MVRangesList (for some methods)}

\item{object}{an MVRangesList (for other methods)}

\item{annotations}{an MVRangesList (for getAnnotations)}

\item{query}{an MVRangesList (for predictCoding)}

\item{subject}{a GRanges, usually}

\item{seqSource}{a BSgenome, usually}

\item{varAllele}{variant alleles}

\item{filterLowQual}{opt. for \code{granges}/\code{summarizeVariants}}

\item{y}{another MVRangesList}
}
\value{
\preformatted{           the MVRangesList
}

\preformatted{         depends on the method invoked.
}
}
\description{
Usually an MVRangesList will be created by callMT.
}
\details{
If the elements were generated by pileupMT(), a coverageRles DataFrame
will be generated and placed into the metadata() list for the result.
Similarly, if an element named 'bam' is located in the metadata of the
supplied arguments, it will be combined into a vector named bamFiles.
}
\section{Utility methods}{


\code{genomeCoverage}       returns estimated mitochondrial read coverage depth
\code{coverage}             returns an RleList of coverage for each sample's chrM
\code{filt}                 removes variants where PASS != TRUE for each element
}

\section{Annotation methods}{


\code{genes}                returns an annotated GRanges of mitochondrial genes
\code{metadata}             get or set elements in the metadata() of an MVRL
\code{getAnnotations}       returns a GRanges of annotated mitochondrial features
\code{genome}               returns the genome (or, perhaps, genomes) in an MVRL
\code{encoding}             returns mutations in coding regions for each element
\code{granges}              returns mildly annotated aggregates of variant sites
\code{snpCall}              retrieves single nucleotide variant polymorphisms
\code{locateVariants}       locates variants within genes, tRNA, rRNA, or D-loop
\code{summarizeVariants}    attempts mass functional annotation of variant sites
\code{consensusString}      creates consensus genotypes from rCRS for eg Haplogrep
}

\section{Visualization methods}{


\code{plot}                 creates circular plot of mitochondrial variant calls
}

