% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileupXenograft.R
\name{pileupXenograft}
\alias{pileupXenograft}
\title{test harness for simultaneous mouse/human MT variant calling (for tracking)}
\usage{
pileupXenograft(bam, sbp = NULL, pup = NULL, callIndels = FALSE,
  minVAF = 0.03, minDepth = 20, verbose = FALSE, ...)
}
\arguments{
\item{bam}{the BAM or object with a @bam slot}

\item{sbp}{optional ScanBamParam object (autogenerated if missing)}

\item{pup}{optional PileupParam object (autogenerated if missing)}

\item{callIndels}{call indels? (This can reduce compute and RAM if FALSE.)}

\item{minVAF}{minimum VAF to accept (default is 0.03, usual NuMT cutoff)}

\item{minDepth}{minimum read depth to accept (default is 20 for sanity)}

\item{verbose}{be verbose, as for profiling? (default is FALSE / quiet)}

\item{...}{additional args to pass on to the pileup() function}
}
\value{
\preformatted{       a list of MVRanges (one human, one mouse)
}
}
\description{
Note that the default pileup params here do NOT distinguish strands in PDXs.
}
\details{
FIXME: add an iterator to create (human, mouse) MVRangesLists when running
pileupXenograft on thousands of individual cells (as in PDX studies).
}
\examples{
\dontrun{
  bam <- "PDX322_TTGCGTCTCCAGAAGG-1.GRCh38_GRCm38_chrM.bam" # in MTseekerData
  mvrl <- pileupXenograft(bam, verbose=TRUE)
}
}
